NEW kraken.py
This is bound to have many errors and such, but this is a version that doesn't have nearly as many as the old one that was given. You can also see what was changed here and on the old file. I've been going line-by-line in the old one to make it work, optimize it and document it.
This is around 10/35ths done